rename : workOn 관련 변수명, 파일명, URL을 attendance로 변경
