+++ src/pages/user/attendance/attendance.module.css
@@ -1,4 +1,4 @@
-.workOnBox {
+.attendanceBox {
   box-sizing: border-box;
   width: 775px;
   padding: 30px 40px;
@@ -34,7 +34,7 @@
   position: absolute;
   right: 0;
 }
-.workOnItem {
+.attendanceItem {
   width: 128px;
   height: 245px;
   display: flex;
@@ -52,7 +52,7 @@
   display: none;
 }
 
-.workOnImg {
+.attendanceImg {
   width: 128px;
   height: 128px;
   border: 3px solid black;
@@ -61,7 +61,7 @@
   display: flex;
 }
 
-.workOnImg img {
+.attendanceImg img {
   width: 100%;
   height: 100%;
   object-fit: cover;
@@ -105,7 +105,7 @@
 }
 
 @media (max-width: 1280px) {
-  .workOnBox {
+  .attendanceBox {
     width: 100%;
     height: 100%;
     border: none;
