+++ src/pages/user/attendance/index.js
@@ -1,0 +1,6 @@
+import { default as render } from './attendanceRender';
+import { default as init } from './attendanceFunc';
+
+const Profile = { render, init };
+
+export default Profile;
