--- src/pages/user/workOn/index.js
@@ -1,6 +1,0 @@
-import { default as render } from './workOnRender';
-import { default as init } from './workOnFunc';
-
-const Profile = { render, init };
-
-export default Profile;
