+++ src/router/routes.js
@@ -9,7 +9,7 @@ import UserProfile from '../pages/user/profile';
 import ProfileForm from '../pages/admin/profileForm';
 import UserProfileForm from '../pages/user/profileForm';
 import WorkCheck from '../pages/user/work';
-import workOn from '../pages/user/workOn';
+import attendance from '../pages/user/attendance';
 
 // 페이지 구조 저장 + 렌더링 페이지 정리
 // async : 비동기 데이터 사용 여부
@@ -24,7 +24,7 @@ const routes = {
   '/admin/notice/view/:noticeSn': { ...Notice.view, async: true, grade: 0 },
   '/admin/notice/insert': { ...Notice.form, async: true, grade: 0 },
   '/admin/notice/update/:noticeSn': { ...Notice.form, async: true, grade: 0 },
-  '/admin/userList': { ...UserList, async:true, grade: 0 },
+  '/admin/userList': { ...UserList, async: true, grade: 0 },
   '/admin/user/:userSn': { ...Profile, grade: 0 },
   '/admin/user/profileForm/:userSn': { ...ProfileForm, grade: 0 },
 
@@ -35,7 +35,7 @@ const routes = {
   '/user/profile': { ...UserProfile, grade: 1 },
   '/user/profile/profileForm': { ...UserProfileForm, grade: 1 },
   '/user/work': { ...WorkCheck, async: true, grade: 1 },
-  '/user/workOn': { ...workOn, grade: 1 },
+  '/user/attendance': { ...attendance, grade: 1 },
 };
 
 export default routes;
